Don't repeat expressions with parameters in order by
Address the failure of the
`ordering.tests.OrderingTests.test_orders_nulls_first_on_filtered_subquery`
test with psycopg3. Because parameters binding happens server side, in a
query such as:

    SELECT DISTINCT a, b, expr(%s)
    ORDRT BY expr(%s) ASC

with params

    [val1, val1]

the parser is unable to identify the match between the ordering
expression and the output column. The query is rewritten as:

    SELECT DISTINCT a, b, expr(%s)
    ORDRT BY 3 ASC

with params

    [val1]

Should any backend not support the numeric alias of the column, the
operation might be guarded by a backend feature (tested with PostgreSQL
and SQLite)