Refs #32948 -- Made WhereNode.clone() use self.__class__().
`Node._new_instance()` was added in 6dd2b54 to work around
`Q.__init__()` having an incompatible signature with `Node.__init__()`.

As `WhereNode` doesn't override `__init__()` we can just use
`self.__class__()` directly which is faster. As an aside, we also do not
need to pass in `children=[]`. That is handled by `Node.__init__()`.