Refs #32948 -- Removed Node._new_instance().
`Node._new_instance()` was added in 6dd2b54 to work around
`Q.__init__()` having an incompatible signature with `Node.__init__()`.

It was intended as a hook that could be overridden if subclasses needed
to change the behaviour of instantiation of their specialised form of
`Node`. In practice this doesn't ever seem to have been used for this
purpose and there are very few calls to `Node._new_instance()` with
other code, e.g. `Node.__deepcopy__()` calling `Node` and overriding
`__class__` as required.